I think I got the SDL widget to draw!
It appears that the 'setAttribute' methods that people online heralded
as the solution to getting QT and SDL to play together were actually
the problem.  Commenting those out allowed SDL to draw to the window.
I'll have to clean up the code later and put in code to update the SDL
surface, but I'm pretty happy to get this working.