Some renaming and some text message work.
In order to keep the naming convention of my classes (which inherit
and override from Qt) consistent with Qt, the symbols in my files are
now in camelCase instead of underscore_style.  I'm not too worried
about Boost or Toast's naming convention since I doubt I'll be
overriding any of those anytime soon. ;)

We also now have the ability to display text messages right on the
canvas!  Right now the text is in the lower-left corner and surrounded
by a blue box, but facilities are available for easily changing those.
Also, the box is dynamically sized to the text and the text is
dynamically sized to the window.  Open up nesT and try it out by dragging
the window around!