We can now render sprites!!!
There were a few fixes to get this working, mostly me accidentally transposing the
sprite data.  There's a few other unrelated fixes that I happend to find, so that's
why so many files are modified.  Also note that we now create the renderer in main()
and then it gets passed to everything that needs it.